implement fu_statusUtil lib
