+++ stats/effects/fu_armoreffects/set_bonuses/nightar/daywalkersetbonuseffect.lua
@@ -1,4 +1,6 @@
 require "/stats/effects/fu_armoreffects/setbonuses_common.lua"
+require "/stats/effects/fu_statusUtil.lua"
+
 setName="fu_daywalkerset"
 
 weaponBonus={
@@ -28,23 +30,6 @@ function init()
 
 	if not world.entityExists(entity.id()) then return end
 	getLight()
-end
-
-function getLight()
-	local position = mcontroller.position()
-	position[1] = math.floor(position[1])
-	position[2] = math.floor(position[2])
-	local lightLevel = math.min(world.lightLevel(position),1.0)
-	lightLevel = math.floor(lightLevel * 100)
-	return lightLevel
-end
-
-function daytimeCheck()
-	return world.timeOfDay() < 0.5 -- true if daytime
-end
-
-function undergroundCheck()
-	return world.underground(mcontroller.position())
 end
 
 function update(dt)
