+++ stats/effects/passive/darkhunternightar.lua
@@ -1,4 +1,5 @@
 require "/scripts/util.lua"
+require "/stats/effects/fu_statusUtil.lua"
 
 function init()
 	script.setUpdateDelta(10)
@@ -14,24 +15,6 @@ function init()
 	self.activateDrainTimer = 0
 	self.didInit=true
 end
-
-function getLight()
-	local position = mcontroller.position()
-	position[1] = math.floor(position[1])
-	position[2] = math.floor(position[2])
-	local lightLevel = math.min(world.lightLevel(position),1.0)
-	lightLevel = math.floor(lightLevel * 100)
-	return lightLevel
-end
-
-function daytimeCheck()
-	return world.timeOfDay() < 0.5 -- true if daytime
-end
-
-function undergroundCheck()
-	return world.underground(mcontroller.position())
-end
-
 
 function update(dt)
 	if not self.didInit then init() end
