+++ stats/effects/passive/nightfear.lua
@@ -1,15 +1,8 @@
+require "/stats/effects/fu_statusUtil.lua"
+
 function init()
 	nightFearEffects=effect.addStatModifierGroup({})
 	script.setUpdateDelta(10)
-end
-
-function getLight()
-	local position = mcontroller.position()
-	position[1] = math.floor(position[1])
-	position[2] = math.floor(position[2])
-	local lightLevel = math.min(world.lightLevel(position),1.0)
-	lightLevel = math.floor(lightLevel * 100)
-	return lightLevel
 end
 
 function update(dt)
