+++ stats/effects/passive/saturnflight.lua
@@ -1,15 +1,4 @@
-function daytimeCheck()
-	return world.timeOfDay() < 0.5 -- true if daytime
-end
-
-function getLight()
-	local position = mcontroller.position()
-	position[1] = math.floor(position[1])
-	position[2] = math.floor(position[2])
-	local lightLevel = math.min(world.lightLevel(position),1.0)
-	lightLevel = math.floor(lightLevel * 100)
-	return lightLevel
-end
+require "/stats/effects/fu_statusUtil.lua"
 
 function update(dt)
 	daytime = daytimeCheck()
