+++ stats/effects/regeneration/darkregen.lua
@@ -1,13 +1,14 @@
 require "/scripts/vec2.lua"
 require "/scripts/util.lua"
 require "/scripts/interp.lua"
+require "/stats/effects/fu_statusUtil.lua"
 
 function init()
 	self.healingRate = 1.01 / config.getParameter("healTime", 420)
 	script.setUpdateDelta(10)
 	self.healTime=config.getParameter("healTime", 180)
+	bonusHandler=effect.addStatModifierGroup({})
 end
-
 
 function activateVisualEffects()
 	local lightLevel = getLight()
@@ -15,21 +16,6 @@ function activateVisualEffects()
 		animator.setParticleEmitterOffsetRegion("blood", mcontroller.boundBox())
 		animator.setParticleEmitterActive("blood", true)
 	end
-end
-
-function getLight()
-	local position = mcontroller.position()
-	position[1] = math.floor(position[1])
-	position[2] = math.floor(position[2])
-	return math.floor(math.min(world.lightLevel(position),1.0) * 100)
-end
-
-function nighttimeCheck()
-	return world.timeOfDay() > 0.5 -- true if daytime
-end
-
-function undergroundCheck()
-	return world.underground(mcontroller.position())
 end
 
 function update(dt)
@@ -42,7 +28,10 @@ function update(dt)
 		else
 			self.healingRate=0.0
 		end
-		--sb.logInfo("darkheal:"..lightLevel..":"..(self.healingRate*dt)..":".. math.max(0,1+status.stat("healingBonus"))..":"..(self.healingRate * dt * math.max(0,1+status.stat("healingBonus"))))
-		status.modifyResourcePercentage("health", self.healingRate * dt * math.max(0,1+status.stat("healingBonus")))
 	end
+	effect.setStatModifierGroup(bonusHandler,{{stat="healthRegen",amount=status.resourceMax("health")*self.healingRate*math.max(0,1+status.stat("healingBonus"))}})
+end
+
+function uninit()
+	effect.removeStatModifierGroup(bonusHandler)
 end
