+++ stats/effects/regeneration/lightregenenergy.lua
@@ -1,27 +1,12 @@
 require "/scripts/vec2.lua"
 require "/scripts/util.lua"
 require "/scripts/interp.lua"
+require "/stats/effects/fu_statusUtil.lua"
 
 function init()
 	self.healingRate = 1.01
 	self.healingTime=config.getParameter("healTime", 320)
 	script.setUpdateDelta(5)
-end
-
-function getLight()
-	local position = mcontroller.position()
-	position[1] = math.floor(position[1])
-	position[2] = math.floor(position[2])
-	return math.floor(math.min(world.lightLevel(position),1.0) * 100)
-end
-
-
-function daytimeCheck()
-	daytime = world.timeOfDay() < 0.5 -- true if daytime
-end
-
-function undergroundCheck()
-	underground = world.underground(mcontroller.position())
 end
 
 function update(dt)
