+++ tech/other/flighttest.lua
@@ -1,5 +1,6 @@
 require "/scripts/vec2.lua"
-local foodThreshold=15
+require "/stats/effects/fu_statusUtil.lua"
+local foodThreshold=15--used by checkFood
 
 function init()
 	if not world.entitySpecies(entity.id()) then return end
@@ -17,10 +18,6 @@ function uninit()
 	animator.setParticleEmitterActive("feathers", false)
 	animator.setParticleEmitterActive("butterflies", false)
 	animator.stopAllSounds("activate")
-end
-
-function checkFood()
-	return (((status.statusProperty("fuFoodTrackerHandler",0)>-1) and status.isResource("food")) and status.resource("food")) or foodThreshold
 end
 
 function activeFlight()
