+++ tech/other/flighttestsaturn.lua
@@ -1,5 +1,6 @@
 require "/scripts/vec2.lua"
-local foodThreshold=15
+require "/stats/effects/fu_statusUtil.lua"
+local foodThreshold=15--used by checkFood
 
 function init()
 	self.active=false
@@ -19,11 +20,6 @@ function uninit()
 	status.removeEphemeralEffect("lowgravflighttech")
 	animator.setParticleEmitterActive("feathers", false)
 end
-
-function checkFood()
-	return (((status.statusProperty("fuFoodTrackerHandler",0)>-1) and status.isResource("food")) and status.resource("food")) or foodThreshold
-end
-
 function boost(direction)
 	self.boostVelocity = vec2.mul(vec2.norm(direction), self.boostSpeed)
 	if self.boostSpeed > 20 then -- prevent super-rapid movement
